[CHORE] Change path TALEARNT-118
