--- src/declareTypes/classNames.d.ts
@@ -1,7 +1,0 @@
-// types/classNames.d.ts
-
-declare module "@utils/classNames" {
-  import { ClassValue } from "clsx";
-
-  export function classNames(...inputs: ClassValue[]): string;
-}
