--- src/pages/auth/SignIn/hook/signIn.api.ts
@@ -1,9 +1,0 @@
-import { postAPI } from "@utils/apiMethods";
-
-import { accountType } from "@pages/auth/common/auth.type";
-import { customAxiosResponseType } from "@type/apiMethods.type";
-
-export const postSignIn = async (
-  account: accountType
-): Promise<customAxiosResponseType<{ accessToken: string } | null>> =>
-  await postAPI("/v1/api/auth/login", account);
