+++ src/pages/auth/api/auth.type.ts
