+++ tsconfig.json
@@ -32,7 +32,6 @@
       "@layout/*": ["src/layout/*"],
       "@pages/*": ["src/pages/*"],
       "@routes/*": ["src/routes/*"],
-      "@type/*": ["src/type/*"],
       "@utils/*": ["src/utils/*"]
     },
     "typeRoots": ["./node_modules/@types", "./src/declareTypes"], // 추가된 types 폴더 경로
